chore(tui): address resize reflow review cleanup
