+++ codex-rs/tui/tests/suite/mod.rs
@@ -1,7 +1,7 @@
 // Aggregates all former standalone integration tests as modules.
 mod model_availability_nux;
 mod no_panic_on_startup;
-mod resize_reflow_smoke;
+mod resize_reflow;
 mod status_indicator;
 mod vt100_history;
 mod vt100_live_commit;
